+++ src/components/index.ts
@@ -9,8 +9,8 @@ export { default as Badge } from './badge';
 export { default as Button } from './button';
 export { default as Box } from './box';
 export { default as Breadcrumbs } from './breadcrumbs';
-export { default as BreadcrumbsItem } from './breadcrumbs/BreadcrumbItem';
-export { default as BreadcrumbLink } from './breadcrumbs/BreadcrumbLink';
+export { default as BreadcrumbsItem } from './breadcrumbs/Item';
+export { default as BreadcrumbLink } from './breadcrumbs/Link';
 export { default as Carousel } from './carousel';
 export { default as Center } from './center';
 export { default as Checkbox } from './checkbox';
